stepps_backend/requirements.txt
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.2
asgiref==3.8.1